fix:separated models
